Move commands under Rails::ConsoleMethods namespace
There's an existing module for the console methods app and helper
so it makes sense to move the commands under this module as commands.rb
already exists for other commands like new, dbconsole, server, etc.
This makes the integration cleaner and easier for other third-party
libraries to add to such as RSpec.